things are looking good

+++ lib/gemology/resque_jobs/extract_metadata.rb
@@ -13,9 +13,8 @@ module Gemology
 
       def initialize( gemfile )
         @gemfile = gemfile
-        @workdir = work_dir_for( gemfile )
+        @workdir = nil # work_dir_for( gemfile )
 
-        logger.info "Starting extraction of metadata from #{@gemfile}"
       end
 
       def work_dir_for( gemfile )
@@ -26,29 +25,36 @@ module Gemology
         return work_dir 
       end
 
-      def save_to_local_file( gemfile ) 
+      def fetch( gemfile )
+        logger.info "Fetching #{gemfile} from cloud container"
+        rubygems_container.object( gemfile )
+      end
+
+      def save_to_local_file( cloud_obj, gemfile ) 
         local_name = File.join( @workdir , gemfile )
-        logger.info "Fetching #{gemfile} from cloud container to #{local_name}"
-        rubygems_container.object( gemfile ).save_to_filename( local_name )
+        logger.info "Saving to #{local_name}"
+        cloud_obj.save_to_filename( local_name )
         return local_name
       end
 
       def run
         begin
-          local_name = save_to_local_file( @gemfile )
-          metadata   = ::Gemology::GemVersionData.new( local_name )
+          logger.info "Starting extraction of metadata from #{@gemfile}"
+          metadata = ::Gemology::GemVersionData.new( fetch( @gemfile ).data )
           ::Gemology::Db.open do |db|
-            db.transaction do |conn|
-              metadata.store_to_db( conn )
-            end
+            metadata.store_to_db( db )
           end
         rescue => e
           logger.error e
           e.backtrace.each { |b| logger.debug b }
           raise e
         ensure
-          logger.info "Cleaning up #{@workdir}"
-          FileUtils.rm_rf @workdir
+
+          if @workdir then
+            logger.info "Cleaning up #{@workdir}"
+            FileUtils.rm_rf @workdir
+          end
+          logger.info "Finished extraction of metadata from #{@gemfile}"
         end
       end
     end
